+++ Ginger Island Mainland Adjustments/Ginger Island Mainland Adjustments/ScheduleManager/IslandSouthPatches.cs
@@ -1,6 +1,7 @@
 ﻿using AtraShared;
 using AtraShared.Utils.Extensions;
 using GingerIslandMainlandAdjustments.AssetManagers;
+using GingerIslandMainlandAdjustments.Configuration;
 using HarmonyLib;
 using Microsoft.Xna.Framework.Graphics;
 using StardewValley.Locations;