+++ Ginger Island Mainland Adjustments/Ginger Island Mainland Adjustments/docs/changelog.md
@@ -18,12 +18,13 @@
 <!-- More schedule debugging tools: get arbitrary schedule from X day? -->
 <!-- Telephone: Lock behind a 10 heart event or something.-->
 <!-- Figure out why Emily dances *in* the changing room? -->
-<!-- IslandNorth and AntiSocial lines for George/Evelyn/Willy-->
+<!-- AntiSocial lines for George/Evelyn/Willy-->
 <!-- Get spouses into Island Farmhouse -->
 <!-- Get Willy to change in his bedroom? -->
 <!-- Make sure the GI schedule keys are right.... -->
 <!-- Document animations -->
 <!-- Consider not allowing spouses to go to Ginger Island if they have a defined marriage_Season_day key?-->
+<!-- Document new tokens -->
 
 ##### Known Issues
 
@@ -35,11 +36,13 @@
 * In theory: integration with Child2NPC.
 * After a certain heart event, you'll can now call Pam, who will tell you if she's headed to Ginger Island. Assuming, that is, you didn't decide to insult her.
 * You can now toggle beach outfits.
+* Adds `GIReturn_Talked_<NPCname>` key, so your spouse can acknowledge that you've talked to them 
 
 Fixes:
 * Schedules are now selected correctly.
 * Fix GILeave/Return dialogue not happening for farmhands. Spouses now have default GILeave/Return lines, although I would recommend still giving your favorite spouse a line or two.
 * `bed` is supported for `GIRemainder` schedules.
+* Fixes for splitscreen.
 
 ### Version 1.1.2